restructure (#177)
