+++ liquidity/ui/src/components/Stats/StatBox.tsx
@@ -44,7 +44,7 @@ export const StatBox = ({ isLoading, title, label, value }: StatBoxProps) => {
         >
           <Fade in>
             <Text fontSize="24px" lineHeight="24px" fontWeight={800}>
-              {value || 'Place'}
+              {value || '—'}
             </Text>
           </Fade>
         </Skeleton>
