+++ liquidity/ui/src/components/Stats/index.ts
@@ -1,1 +1,2 @@
 export * from './StatBox';
+export * from './StatsList';
